Rename .h files and start adding Item's children
Renamed .h files to .hpp as well as added a .gitattributes to normalize line endings
