+++ Magic.hpp
@@ -1,5 +1,5 @@
-#ifndef MAGIC_H
-#define MAGIC_H
+#ifndef MAGIC_HPP
+#define MAGIC_HPP
 class Magic{
 public:
     Magic();
