--- Map.h
@@ -1,7 +1,0 @@
-#ifndef MAP_H
-#define MAP_H
-class Map{
-public:
-  Map();
-};
-#endif
