+++ Map.hpp
@@ -1,0 +1,7 @@
+#ifndef MAP_HPP
+#define MAP_HPP
+class Map{
+public:
+    Map();
+};
+#endif
